javadoc additions and related refactoring...
