+++ .gitignore
@@ -61,3 +61,6 @@ com_crashlytics_export_strings.xml
 crashlytics.properties
 crashlytics-build.properties
 
+# project-specific stuff
+apidocs
+
