+++ modules/core/src/main/java/io/swagger/assert4j/assertions/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Assertion builders and utilities
+ */
+package io.swagger.assert4j.assertions;
