+++ modules/core/src/main/java/io/swagger/assert4j/attachments/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Attachment builder and utilities
+ */
+package io.swagger.assert4j.attachments;
