+++ modules/core/src/main/java/io/swagger/assert4j/auth/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Builders and utilities for HTTP Authentication functionality
+ */
+package io.swagger.assert4j.auth;
