+++ modules/core/src/main/java/io/swagger/assert4j/execution/RecipeExecutor.java
@@ -1,6 +1,5 @@
 package io.swagger.assert4j.execution;
 
-import io.swagger.assert4j.ExecutionListener;
 import io.swagger.assert4j.TestRecipe;
 
 import java.util.List;
