+++ modules/core/src/main/java/io/swagger/assert4j/execution/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * TestRecipe Execution functionality
+ */
+package io.swagger.assert4j.execution;
