+++ modules/core/src/main/java/io/swagger/assert4j/extractor/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Extractor classes and utilities
+ */
+package io.swagger.assert4j.extractor;
