+++ modules/core/src/main/java/io/swagger/assert4j/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * Root package containing core TestRecipe functionality
+ */
+
+package io.swagger.assert4j;
