+++ modules/core/src/main/java/io/swagger/assert4j/properties/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Property-related classes and utilities.
+ */
+package io.swagger.assert4j.properties;
