+++ modules/core/src/main/java/io/swagger/assert4j/result/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Classes for handling execution results
+ */
+package io.swagger.assert4j.result;
