+++ modules/core/src/main/java/io/swagger/assert4j/support/ExecutionLogger.java
@@ -4,8 +4,8 @@ import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
 import io.swagger.assert4j.client.model.ProjectResultReport;
 import io.swagger.assert4j.client.model.TestSuiteResultReport;
-import io.swagger.assert4j.ExecutionListener;
 import io.swagger.assert4j.execution.Execution;
+import io.swagger.assert4j.execution.ExecutionListener;
 import io.swagger.assert4j.result.TestStepResult;
 import io.swagger.util.Json;
 import org.apache.commons.lang3.StringUtils;
