+++ modules/core/src/main/java/io/swagger/assert4j/support/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Common utilities
+ */
+package io.swagger.assert4j.support;
