+++ modules/core/src/main/java/io/swagger/assert4j/teststeps/delay/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Delay TestStep classes
+ */
+package io.swagger.assert4j.teststeps.delay;
