+++ modules/core/src/main/java/io/swagger/assert4j/teststeps/groovyscript/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Groovy TestStep classes
+ */
+package io.swagger.assert4j.teststeps.groovyscript;
