+++ modules/core/src/main/java/io/swagger/assert4j/teststeps/jdbcrequest/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * JDBC TestStep classes
+ */
+package io.swagger.assert4j.teststeps.jdbcrequest;
