+++ modules/core/src/main/java/io/swagger/assert4j/teststeps/mockresponse/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * SOAP MockResponse TestStep classes
+ */
+package io.swagger.assert4j.teststeps.mockresponse;
