+++ modules/core/src/main/java/io/swagger/assert4j/teststeps/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Base TestStep definitions
+ */
+package io.swagger.assert4j.teststeps;
