+++ modules/core/src/main/java/io/swagger/assert4j/teststeps/plugin/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Plugin TestStep classes
+ */
+package io.swagger.assert4j.teststeps.plugin;
