+++ modules/core/src/main/java/io/swagger/assert4j/teststeps/properties/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Property TestStep classes
+ */
+package io.swagger.assert4j.teststeps.properties;
