+++ modules/core/src/main/java/io/swagger/assert4j/teststeps/propertytransfer/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * PropertyTransfer TestStep classes
+ */
+package io.swagger.assert4j.teststeps.propertytransfer;
