+++ modules/core/src/main/java/io/swagger/assert4j/teststeps/request/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Base HTTP TestStep classes
+ */
+package io.swagger.assert4j.teststeps.request;
