+++ modules/core/src/main/java/io/swagger/assert4j/teststeps/restrequest/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * REST TestStep classes
+ */
+package io.swagger.assert4j.teststeps.restrequest;
