+++ modules/core/src/main/java/io/swagger/assert4j/teststeps/soaprequest/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * SOAP TestStep classes
+ */
+package io.swagger.assert4j.teststeps.soaprequest;
