+++ modules/core/src/main/java/overview.html
@@ -1,0 +1,5 @@
+<html>
+<body>
+Swagger Assert4j JavaDocs - check out the <a href="https://github.com/SmartBear/swagger-assert4j" target="_new">documentation on GitHub</a> to get started!
+</body>
+</html>
