+++ modules/cucumber/modules/core/src/main/java/io/swagger/assert4j/cucumber/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Core cucumber execution classes
+ */
+package io.swagger.assert4j.cucumber;
