+++ modules/cucumber/modules/runner/src/main/java/io/swagger/assert4j/cucumber/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Cucumber CLI classes
+ */
+package io.swagger.assert4j.cucumber;
