+++ modules/cucumber/modules/samples/src/main/java/com/smartbear/cucumber/samples/extension/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Cucumber sample classes
+ */
+package com.smartbear.cucumber.samples.extension;
