+++ modules/cucumber/modules/stepdefs/src/main/java/io/swagger/assert4j/cucumber/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Cucumber StepDefs for REST / Swagger API testing
+ */
+package io.swagger.assert4j.cucumber;
