+++ modules/facade/src/main/java/io/swagger/assert4j/facade/execution/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Recipe execution utilities
+ */
+package io.swagger.assert4j.facade.execution;
