+++ modules/local/src/main/java/io/swagger/assert4j/local/execution/SoapUIRecipeExecutor.java
@@ -11,17 +11,17 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.collect.Lists;
 import com.smartbear.ready.recipe.JsonRecipeParser;
 import com.smartbear.ready.recipe.teststeps.TestCaseStruct;
+import io.swagger.assert4j.TestRecipe;
 import io.swagger.assert4j.client.model.ProjectResultReport;
 import io.swagger.assert4j.client.model.TestStep;
-import io.swagger.assert4j.ExecutionListener;
-import io.swagger.assert4j.TestRecipe;
-import io.swagger.assert4j.execution.DataExtractors;
 import io.swagger.assert4j.execution.Execution;
+import io.swagger.assert4j.execution.ExecutionListener;
 import io.swagger.assert4j.execution.ExecutionMode;
 import io.swagger.assert4j.execution.RecipeExecutionException;
 import io.swagger.assert4j.execution.RecipeExecutor;
 import io.swagger.assert4j.execution.RecipeFilter;
 import io.swagger.assert4j.execution.UnsupportedTestStepException;
+import io.swagger.assert4j.extractor.DataExtractors;
 import io.swagger.assert4j.teststeps.TestStepTypes;
 
 import java.util.Arrays;
