+++ modules/local/src/main/java/io/swagger/assert4j/local/execution/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Local recipe execution classes
+ */
+package io.swagger.assert4j.local.execution;
