+++ modules/local/src/test/java/io/swagger/assert4j/local/execution/SoapUIRecipeExecutorTest.java
@@ -1,25 +1,19 @@
 package io.swagger.assert4j.local.execution;
 
 import com.google.gson.Gson;
+import io.swagger.assert4j.TestRecipe;
 import io.swagger.assert4j.client.model.HarResponse;
 import io.swagger.assert4j.client.model.ProjectResultReport;
+import io.swagger.assert4j.execution.Execution;
+import io.swagger.assert4j.execution.ExecutionListener;
+import io.swagger.assert4j.extractor.ExtractorData;
+import io.swagger.assert4j.teststeps.propertytransfer.PathLanguage;
 import io.swagger.assert4j.util.rest.JsonTestObject;
 import io.swagger.assert4j.util.rest.Pair;
-import io.swagger.assert4j.ExecutionListener;
-import io.swagger.assert4j.TestRecipe;
-import io.swagger.assert4j.execution.Execution;
-import io.swagger.assert4j.extractor.ExtractorData;
-import io.swagger.assert4j.local.execution.SoapUIRecipeExecutor;
-import io.swagger.assert4j.teststeps.propertytransfer.PathLanguage;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.Test;
 
-import static io.swagger.assert4j.util.rest.local.LocalServerUtil.addGetToLocalServer;
-import static io.swagger.assert4j.util.rest.local.LocalServerUtil.addPostToLocalServer;
-import static io.swagger.assert4j.util.rest.local.LocalServerUtil.getPostedJsonTestObject;
-import static io.swagger.assert4j.util.rest.local.LocalServerUtil.startLocalServer;
-import static io.swagger.assert4j.util.rest.local.LocalServerUtil.stopLocalServer;
 import static io.swagger.assert4j.TestRecipeBuilder.newTestRecipe;
 import static io.swagger.assert4j.extractor.Extractors.fromProperty;
 import static io.swagger.assert4j.teststeps.TestSteps.GET;
@@ -29,6 +23,11 @@ import static io.swagger.assert4j.teststeps.TestSteps.propertyTransfer;
 import static io.swagger.assert4j.teststeps.propertytransfer.PropertyTransferBuilder.from;
 import static io.swagger.assert4j.teststeps.propertytransfer.PropertyTransferSourceBuilder.aSource;
 import static io.swagger.assert4j.teststeps.propertytransfer.PropertyTransferTargetBuilder.aTarget;
+import static io.swagger.assert4j.util.rest.local.LocalServerUtil.addGetToLocalServer;
+import static io.swagger.assert4j.util.rest.local.LocalServerUtil.addPostToLocalServer;
+import static io.swagger.assert4j.util.rest.local.LocalServerUtil.getPostedJsonTestObject;
+import static io.swagger.assert4j.util.rest.local.LocalServerUtil.startLocalServer;
+import static io.swagger.assert4j.util.rest.local.LocalServerUtil.stopLocalServer;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.CoreMatchers.not;
 import static org.hamcrest.CoreMatchers.nullValue;
