+++ modules/maven-plugin/src/main/java/io/swagger/assert4j/maven/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Maven plugin classes
+ */
+package io.swagger.assert4j.maven;
