+++ modules/oas/src/main/java/io/swagger/assert4j/swagger/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Swagger-related extensions to core API
+ */
+package io.swagger.assert4j.swagger;
