+++ modules/testserver/src/main/java/io/swagger/assert4j/testserver/execution/AbstractTestServerExecutor.java
@@ -5,9 +5,9 @@ import io.swagger.assert4j.client.model.RequestTestStepBase;
 import io.swagger.assert4j.client.model.TestCase;
 import io.swagger.assert4j.client.model.TestStep;
 import io.swagger.assert4j.client.model.UnresolvedFile;
-import io.swagger.assert4j.ExecutionListener;
-import io.swagger.assert4j.execution.DataExtractors;
 import io.swagger.assert4j.execution.Execution;
+import io.swagger.assert4j.execution.ExecutionListener;
+import io.swagger.assert4j.extractor.DataExtractors;
 import io.swagger.assert4j.extractor.ExtractorData;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
