+++ modules/testserver/src/main/java/io/swagger/assert4j/testserver/execution/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Classes for accessing TestServer functionality
+ */
+package io.swagger.assert4j.testserver.execution;
