+++ modules/testserver/src/main/java/io/swagger/assert4j/testserver/teststeps/datasource/datagen/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Data generator classes
+ */
+package io.swagger.assert4j.testserver.teststeps.datasource.datagen;
