+++ modules/testserver/src/main/java/io/swagger/assert4j/testserver/teststeps/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Server TestSteps utilities
+ */
+package io.swagger.assert4j.testserver.teststeps;
